+++ lib/src/mock_navigator.dart
@@ -146,6 +146,11 @@ class _MockNavigatorState extends NavigatorState {
   }
 
   @override
+  Future<bool> maybePop<T extends Object?>([T? result]) {
+    return _navigator.maybePop<T>(result);
+  }
+
+  @override
   Future<T?> pushAndRemoveUntil<T extends Object?>(
     Route<T> newRoute,
     RoutePredicate predicate,