+++ test/src/mock_navigator_test.dart
@@ -186,6 +186,22 @@ void main() {
       verify(() => navigator.popUntil(testRoutePredicate)).called(1);
     });
 
+    testWidgets('mocks .maybePop calls', (tester) async {
+      when(() => navigator.maybePop(any<dynamic>()))
+          .thenAnswer((_) async => true);
+
+      await tester.pumpTest(
+        navigator: navigator,
+        builder: (context) => TextButton(
+          onPressed: () => Navigator.of(context).maybePop(testRoutePredicate),
+          child: const Text('Trigger'),
+        ),
+      );
+
+      await tester.tap(find.byType(TextButton));
+      verify(() => navigator.maybePop(testRoutePredicate)).called(1);
+    });
+
     testWidgets('mocks .pushAndRemoveUntil calls', (tester) async {
       when(() => navigator.pushAndRemoveUntil(any(), any()))
           .thenAnswer((_) async {});